Improve migration IT
